[bugfix] [qa] Restored the ability to run tests.
This is not pretty. `test_fn` is no longer defined in a `mod test` inside `test.rs`,
because that makes it inaccessible from the outside as `egg::test_fn!`.
It can still be accessed as `test_fn!` but eww.
I tried every suggestion from SO. Scoping rules for macros are just weird.